feat(test): add more test for <WriteButton>; move ERROR_MESSAGES to enums;
