+++ src/common/utils/form/error.ts
@@ -1,8 +1,8 @@
 import { ApolloError } from 'apollo-client'
 import { MessageDescriptor } from 'react-intl'
 
-import { ERROR_CODES } from '~/common/enums'
-import { ERROR_MESSAGES, getErrorCodes } from '~/components/GQL'
+import { ERROR_CODES, ERROR_MESSAGES } from '~/common/enums'
+import { getErrorCodes } from '~/components/GQL'
 
 type ErrorMessages = { [key in ERROR_CODES]: MessageDescriptor }
