+++ src/components/ArticleDigest/DropdownActions/index.tsx
@@ -3,7 +3,7 @@ import _pickBy from 'lodash/pickBy'
 import { useContext } from 'react'
 import { FormattedMessage } from 'react-intl'
 
-import { ERROR_CODES } from '~/common/enums'
+import { ERROR_CODES, ERROR_MESSAGES } from '~/common/enums'
 import { translate } from '~/common/utils'
 import {
   AddCollectionsArticleDialog,
@@ -11,7 +11,6 @@ import {
   BookmarkButton,
   Button,
   Dropdown,
-  ERROR_MESSAGES,
   FingerprintDialog,
   IconMore16,
   IconSize,
