+++ src/components/Buttons/Write/Write.test.tsx
@@ -1,7 +1,7 @@
 import mockRouter from 'next-router-mock'
 import { describe, expect, it } from 'vitest'
 
-import { PATHS, TEST_ID } from '~/common/enums'
+import { ERROR_CODES, ERROR_MESSAGES, PATHS, TEST_ID } from '~/common/enums'
 import { render, screen } from '~/common/utils/test'
 import { WriteButton } from '~/components'
 
@@ -31,5 +31,18 @@ describe('<WriteButton>', () => {
     button.click()
 
     expect(screen.getByTestId(TEST_ID.DIALOG_LIKECOIN)).toBeInTheDocument()
+  })
+
+  it('should show toast when forbidden', () => {
+    render(<WriteButton allowed authed forbidden />)
+
+    const button = screen.getByRole('button', { name: 'Create' })
+    button.click()
+
+    const $toast = screen.getByRole('alert')
+    expect($toast).toBeInTheDocument()
+    expect($toast).toHaveTextContent(
+      ERROR_MESSAGES[ERROR_CODES.FORBIDDEN_BY_STATE].defaultMessage as string
+    )
   })
 })
