+++ src/components/Comment/DropdownActions/index.tsx
@@ -4,14 +4,13 @@ import _pickBy from 'lodash/pickBy'
 import { useContext } from 'react'
 import { FormattedMessage } from 'react-intl'
 
-import { ERROR_CODES } from '~/common/enums'
+import { ERROR_CODES, ERROR_MESSAGES } from '~/common/enums'
 import { translate } from '~/common/utils'
 import {
   Button,
   CommentFormDialog,
   CommentFormType,
   Dropdown,
-  ERROR_MESSAGES,
   IconMore16,
   LanguageContext,
   Menu,
