+++ src/components/Comment/FooterActions/index.tsx
@@ -2,11 +2,14 @@ import gql from 'graphql-tag'
 import { useContext } from 'react'
 import { FormattedMessage } from 'react-intl'
 
-import { ERROR_CODES, OPEN_LIKE_COIN_DIALOG } from '~/common/enums'
+import {
+  ERROR_CODES,
+  ERROR_MESSAGES,
+  OPEN_LIKE_COIN_DIALOG,
+} from '~/common/enums'
 import { translate } from '~/common/utils'
 import {
   CommentFormType,
-  ERROR_MESSAGES,
   LanguageContext,
   toast,
   ViewerContext,
