+++ src/views/ArticleDetail/SupportWidget/SupportButton/index.tsx
@@ -3,17 +3,12 @@ import { FormattedMessage } from 'react-intl'
 
 import {
   ERROR_CODES,
+  ERROR_MESSAGES,
   OPEN_UNIVERSAL_AUTH_DIALOG,
   REFETCH_DONATORS,
 } from '~/common/enums'
 import { analytics } from '~/common/utils'
-import {
-  DonationDialog,
-  ERROR_MESSAGES,
-  toast,
-  Translate,
-  ViewerContext,
-} from '~/components'
+import { DonationDialog, toast, Translate, ViewerContext } from '~/components'
 import DonationButton from '~/components/Buttons/DonationButton'
 import {
   ArticleDetailPublicQuery,
