+++ src/views/ArticleDetail/Toolbar/CommentBar/index.tsx
@@ -5,6 +5,7 @@ import { FormattedMessage } from 'react-intl'
 import {
   CLOSE_ACTIVE_DIALOG,
   ERROR_CODES,
+  ERROR_MESSAGES,
   OPEN_LIKE_COIN_DIALOG,
   OPEN_UNIVERSAL_AUTH_DIALOG,
   PATHS,
@@ -17,7 +18,6 @@ import {
   Card,
   CardProps,
   CommentFormDialog,
-  ERROR_MESSAGES,
   IconComment16,
   LanguageContext,
   Media,
