+++ src/views/ArticleDetail/Toolbar/DonationButton/index.tsx
@@ -2,12 +2,16 @@ import gql from 'graphql-tag'
 import { useContext } from 'react'
 import { FormattedMessage } from 'react-intl'
 
-import { ERROR_CODES, OPEN_UNIVERSAL_AUTH_DIALOG, TEXT } from '~/common/enums'
+import {
+  ERROR_CODES,
+  ERROR_MESSAGES,
+  OPEN_UNIVERSAL_AUTH_DIALOG,
+  TEXT,
+} from '~/common/enums'
 import { analytics, numAbbr, translate } from '~/common/utils'
 import {
   Button,
   DonationDialog,
-  ERROR_MESSAGES,
   IconDonate24,
   LanguageContext,
   TextIcon,
